+++ src/client/components/NewPassword/NewPassword.test.tsx
@@ -1,17 +1,17 @@
-import { screen, render, fireEvent } from '@testing-library/react';
+import { screen, fireEvent } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
-import { MemoryRouter } from 'react-router-dom';
+import { renderWithProviders } from '../../__tests__/test-utils';
 import NewPassword from './NewPassword';
 
 describe('New Password submit form', () => {
   test('should render an empty input field with a New Password label', () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const newPassword = screen.getByLabelText('New Password');
     expect(newPassword).toHaveValue('');
   });
 
   test('user input should change the new password field', () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const newPassword = screen.getByLabelText('New Password');
     const newInput = 'myNewPassword';
     fireEvent.change(newPassword, {target: {value: newInput}});
@@ -19,13 +19,13 @@ describe('New Password submit form', () => {
   });
 
   test('should render an empty input field with a Confirm Password label', () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const confirmPassword = screen.getByLabelText('Confirm Password');
     expect(confirmPassword).toHaveValue('');
   });
 
   test('user input should change the confirm new password field', () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const confirmPassword = screen.getByLabelText('Confirm Password');
     const newInput = 'confirmedPassword';
     fireEvent.change(confirmPassword, {target: {value: newInput}});
@@ -33,13 +33,13 @@ describe('New Password submit form', () => {
   });
 
   test('should render a button to submit a new password', () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const button = screen.getByRole('button', {name: 'Change Password'});
     expect(button).toBeVisible();
   });
 
   test('user submitting passwords that do not match causes and error', async () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const user = userEvent.setup();
     const newPassword = screen.getByLabelText('New Password');
     const confirmPassword = screen.getByLabelText('Confirm Password');
@@ -52,7 +52,7 @@ describe('New Password submit form', () => {
   });
 
   test('user submitting a password that does not pass all tests causes error', async () => {
-    render(<MemoryRouter><NewPassword/></MemoryRouter>);
+    renderWithProviders(<NewPassword/>);
     const user = userEvent.setup();
     const newPassword = screen.getByLabelText('New Password');
     const confirmPassword = screen.getByLabelText('Confirm Password');
