+++ src/client/components/NewPassword/NewPassword.tsx
