+++ src/client/components/NewPassword/index.ts
