+++ src/client/components/PrivateRoute/index.ts
