+++ src/client/components/Signup/Signup.test.tsx
@@ -1,6 +1,6 @@
-import { render, screen, fireEvent } from '@testing-library/react';
+import { screen, fireEvent } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { MemoryRouter } from 'react-router-dom';
+import { renderWithProviders } from '../../__tests__/test-utils';
 import Signup from './Signup';
 
 describe('Signup form', () => {
@@ -12,13 +12,13 @@ describe('Signup form', () => {
   ];
 
   test('input field with username label renders empty', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const usernameInput = screen.getByLabelText('Username');
     expect(usernameInput).toHaveValue('');
   });
 
   test('user input changes value of username input field', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const usernameInput = screen.getByRole('textbox', {name: 'Username'});
     const newInput = 'myUsername';
     fireEvent.change(usernameInput, {target: {value: newInput}});
@@ -26,13 +26,13 @@ describe('Signup form', () => {
   });
 
   test('input field with password label renders empty', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const passwordInput = screen.getByLabelText('Password');
     expect(passwordInput).toHaveValue('');
   });
 
   test('user clicking on password input field should display tooltip', async () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const user = userEvent.setup();
     const passwordInput = screen.getByLabelText('Password');
     await user.click(passwordInput);
@@ -41,7 +41,7 @@ describe('Signup form', () => {
   });
 
   test('user input changes value of password input field', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const passwordInput = screen.getByLabelText('Password');
     const newInput = 'myPassword';
     fireEvent.change(passwordInput, {target: {value : newInput}});
@@ -49,13 +49,13 @@ describe('Signup form', () => {
   });
 
   test('input field with confirm password label renders empty', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const confirmPassword = screen.getByLabelText('Confirm Password');
     expect(confirmPassword).toHaveValue('');
   });
 
   test('user focusing on confirm password input field should display tooltip', async () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const user = userEvent.setup();
     const confirmPassword = screen.getByLabelText('Confirm Password');
     await user.click(confirmPassword);
@@ -64,7 +64,7 @@ describe('Signup form', () => {
   });
 
   test('user input changes value of confirm password input field', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const confirmPassword = screen.getByLabelText('Confirm Password');
     const newInput = 'confirmPassword';
     fireEvent.change(confirmPassword, {target: {value: newInput}});
@@ -72,19 +72,19 @@ describe('Signup form', () => {
   });
 
   test('select field with security questions label renders with correct default option', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const question = screen.getByLabelText('Security Question');
     expect(question).toHaveValue(options[0]);
   });
 
   test('select field renders with correct number of options', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const questions = screen.getAllByRole('option');
     expect(questions.length).toBe(options.length);
   });
 
   test('user input changes value of security question select field', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const question = screen.getByRole('combobox', {name: 'Security Question'});
     fireEvent.change(question, {target: {value: options[2]}});
     const option = (screen.getByRole('option', {name: options[2]})) as HTMLOptionElement;
@@ -94,13 +94,13 @@ describe('Signup form', () => {
   });
 
   test('input field with answer label renders empty', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const answer = screen.getByLabelText('Answer');
     expect(answer).toHaveValue('');
   });
 
   test('user input changes value of answer input field', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const answer = screen.getByRole('textbox', {name: 'Answer'});
     const newInput = 'myAnswer';
     fireEvent.change(answer, {target: {value: newInput}});
@@ -110,7 +110,7 @@ describe('Signup form', () => {
   test('user not picking a security question but filling all other' +
     'fields correctly and clicking the "Sign up" button should display' +
     'an error message', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+      renderWithProviders(<Signup/>);
     const username = screen.getByRole('textbox', {name: 'Username'});
     fireEvent.change(username, {target: {value: 'user'}});
     const password = screen.getByLabelText('Password');
@@ -126,13 +126,13 @@ describe('Signup form', () => {
   });
 
   test('renders a signup button', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const signup = screen.getByRole('button', {name: 'Sign up'});
     expect(signup).toBeVisible();
   })
 
   test('renders a link to go back to login page', () => {
-    render(<MemoryRouter><Signup/></MemoryRouter>);
+    renderWithProviders(<Signup/>);
     const loginLink = screen.getByRole('link', {name: 'Login'});
     expect(loginLink).toBeVisible();
   });
