+++ vite.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: './src/__tests__/setup.ts'
+    setupFiles: './src/client/__tests__/setup.ts'
   },
   server: {
     proxy: {
